Stopping at this point as I have gone over time. I started by implementing an interface calculates the discount, with a default method that gives 0 discount. I wasted some time going down a discount service route which didnt really go anywhere.
Instead each Unit item can calculate the discount to be applied. Each product has a discount and a product code to identify it. The discount is any object that implements the ItemDiscount interface to allow for multiple different implementations. I can then read through all the items in the basket and apply any discounts required. This didnt quite work out as i intended, as my stream wasnt catching only the distinct items as I intended it to.